Fix OSD boundary exceeding the frame boundary
Using the v4l2 command the OSD feature can be disabled, but actually
it is not disabled in the hardware, instead it is shifted to out
of resolution boundary(frame boundary), so that it won't appear on
the screen. But when the resoution is increased, the OSD is appered
back on the screen though it is disabled.
To fix this a separate register is added so that the OSD will get disabled in the hardware itself.

Libro verion update in the tcl

Signed-off-by: Shravan Chippa <shravan.chippa@microchip.com>